Move lib modules to only be useable on the server
