+++ app/components/Library/Bandcamp.tsx
@@ -2,7 +2,7 @@ import clsx from 'clsx'
 
 import Card from '~/components/Base/Card'
 import { A, Button, ButtonLink } from '~/components/Base'
-import type { SavedBandcampItem } from '~/lib/types/library'
+import type { SavedBandcampItem } from '~/lib/types/library.server'
 import useLibrary from '~/hooks/useLibrary'
 
 const searchParams = new URLSearchParams({
