+++ app/routes/genre/random.tsx
@@ -1,6 +1,6 @@
 import { redirect } from '@remix-run/node'
 
-import db from '~/lib/db'
+import db from '~/lib/db.server'
 
 export async function loader() {
   const genre = await db.getRandomGenre()
