+++ app/routes/genres.tsx
@@ -2,7 +2,7 @@ import { json } from '@remix-run/node'
 import { useLoaderData } from '@remix-run/react'
 import clsx from 'clsx'
 
-import db from '~/lib/db'
+import db from '~/lib/db.server'
 import AlbumErrorBoundary from '~/components/Album/ErrorBoundary'
 import { Layout, Heading, Container } from '~/components/Base'
 import GenreSearchForm from '~/components/Forms/GenreSearch'
