+++ app/routes/group/$slug.tsx
@@ -1,7 +1,7 @@
 import { useLoaderData } from '@remix-run/react'
 import { json, LoaderArgs } from '@remix-run/node'
 
-import spotifyLib from '~/lib/spotify'
+import spotifyLib from '~/lib/spotify.server'
 import Album from '~/components/Album'
 import AlbumErrorBoundary from '~/components/Album/ErrorBoundary'
 import { Layout } from '~/components/Base'
