+++ app/routes/index.tsx
@@ -3,9 +3,9 @@ import { json, LoaderArgs } from '@remix-run/node'
 import { useLoaderData } from '@remix-run/react'
 import clsx from 'clsx'
 
-import auth from '~/lib/auth'
-import db from '~/lib/db'
-import spotifyLib from '~/lib/spotify'
+import auth from '~/lib/auth.server'
+import db from '~/lib/db.server'
+import spotifyLib from '~/lib/spotify.server'
 import { Heading, Layout, Container, Link, ButtonLink } from '~/components/Base'
 import RelatedArtistSearchForm from '~/components/Forms/RelatedArtistSearch'
 import GenreSearchForm from '~/components/Forms/GenreSearch'
