+++ app/routes/random.tsx
@@ -1,7 +1,7 @@
 import { redirect, LoaderArgs } from '@remix-run/node'
 
 import random from '~/lib/random.server'
-import spotifyLib from '~/lib/spotify'
+import spotifyLib from '~/lib/spotify.server'
 
 export async function loader({ request }: LoaderArgs) {
   const spotify = await spotifyLib.initializeFromRequest(request)
