+++ app/routes/spotify/featured-playlist.tsx
@@ -1,7 +1,7 @@
 import { json, LoaderArgs } from '@remix-run/node'
 import { useLoaderData } from '@remix-run/react'
 
-import spotifyLib from '~/lib/spotify'
+import spotifyLib from '~/lib/spotify.server'
 import { Layout } from '~/components/Base'
 import Playlist from '~/components/Album/Playlist'
 import AlbumErrorBoundary from '~/components/Album/ErrorBoundary'
