+++ app/routes/spotify/new-releases.tsx
@@ -1,7 +1,7 @@
 import { LoaderArgs, json } from '@remix-run/node'
 import { useLoaderData } from '@remix-run/react'
 
-import spotifyLib from '~/lib/spotify'
+import spotifyLib from '~/lib/spotify.server'
 import { Layout } from '~/components/Base'
 import Album from '~/components/Album'
 import AlbumErrorBoundary from '~/components/Album/ErrorBoundary'
