+++ app/routes/twitter/$username.tsx
@@ -2,8 +2,8 @@ import { LoaderArgs, json } from '@remix-run/node'
 import { useLoaderData } from '@remix-run/react'
 import clsx from 'clsx'
 
-import db from '~/lib/db'
-import spotifyLib from '~/lib/spotify'
+import db from '~/lib/db.server'
+import spotifyLib from '~/lib/spotify.server'
 import { Layout, A } from '~/components/Base'
 import Album from '~/components/Album'
 import Playlist from '~/components/Album/Playlist'
